Port Faraway Island from Emerald
Faraway Island, the Mew hide-and-seek game, some tiles, and its music has been imported from Emerald. The Vermillion dockmaster has new multichoices to facilitate getting to the island. There are no means to receive the Old Sea Map, but the item list has been expanded to include it. This commit also adds all the map section ids from Emerald, so that Firered could display and write the Mew's met location properly. The Mew is special-cased to have Emerald written as its game origin so that it is more legal. This does not affect its obedience.

The two main functional differences is that the player can't bike on the island (the long grass doesn't stop the bike like in RSE), and the interior of the island doesn't have the shade weather (Mew fadescreens during its script, which doubles up and permanently darkens the OW more than normal with the shade weather.) Aesthetically, the island is a mix of Firered and Emerald tiles, with Emerald's grass tiles and trees imported.