Remove Linking Restrictions
Can trade with Ruby, Sapphire, and (completed) Emerald from the start. Only tested link trades with Sapphire. It should be okay in the Union Room as well, but I can't test that.

Should also be able to link with Colosseum and XD from the start, but that hasn't been tested.

To facilitate trading with XD, FLAG_SYS_GAME_CLEAR has been moved to another flag, while the original is now named FLAG_SYS_GAME_CLEAR_OLD and enabled the moment the player gets the Pokedex.

Additonally, Pokemon that evolve by trade with a held item will evolve correctly without the National Dex. This was overlooked in the last commit.