Implement Key System Menu
The menu for the Key System is now implemented and is accessed on the main menu (where Continue and New Game are). The Help System has been edited to explain the keys and is always available while in this menu.

Only the difficulty mode is actually working in game. The other values aren't hooked up to anything.