Implement Wild Encounters For Version Switch
Wild Encounters now switch appropriately. The Pokedex's area function also only shows the areas for the current version.

Additionally, fixed a bug where the save would be reloaded upon exiting the titlescreen and finishing the quest log scenes, which would wipe changed keys. The keys are now backed up and restored when the save is reloaded in those spots.