Implement IV and EV Calc Keys
The IV and EV Calc keys are in and working. They should only affect the player's party with the exception of the All Perfect setting for IVs also affecting enemy trainers. The three enemy trainers with Hidden Power on a mon still have custom IVs for that particular mon when All Perfect is set.

Recalculating the party's stats on save reload and a mon's stats when caught and sent to the party results in some weird changes in current HP. It does not revive fainted Pokemon nor set Pokemon to negative HP though, so it is staying for now.