Pick Version on New Game & Random Titlescreen if Empty Save
The game will now prompt the player for which version they want to play on before starting a new game.

There are two different ways to trigger this because I did one way first and discovered that it froze when trying to start a new game automatically (because of an empty or invalid save file). I'm not taking the time to pare it down to one.

The titlescreen version is now chosen at random if the save is empty, invalid, or the 1M-circuit board isn't installed. It's a rougly 50% chance for either.