Implement No Free Heals Key & Prepare For Nuzlocke Implementation
Created a new key that disables free heals from HealPlayerParty, most notably from Pokemon Centers. The PC now stores a Pokemon's HP and status condition when this key is set, so it is no longer a free heal. Caught Pokemon will also not be healed when sent to the box.

In the process of implementing this key, made Nuzlocke mode not heal fainted Pokemon from HealPlayerParty, and the PC box will store a Pokemon's HP if it is fainted, preventing it from being used as a Revive.

For both modes, heals are fully done if linking or right after the first battle.

Additionally, fixed a missed instance (entering the Union Room) where the linking flag should be set and the party's stats recalculated.