Prevent Dupe Clause Circumvention & Add 1st Encounter Indicator
Separate dex flags are now stored to keep track of captures in Nuzlocke mode. Fainted captures do not set these flags. These flags are now checked when determining dupes instead of the standard owned flags, preventing players from cheating the system by catching lots of fainted Pokemon elsewhere (to increase the likelihood of dupes) and forcing a favorable 1st encounter. Now only usable caught Pokemon will count for the dupe clause.

There is now an indicator for when a Pokemon is the first encounter in an area. It is a red 1 printed where the Pokeball symbol is and always overwrites it.

Additionally, no Nuzlocke area flags are set until the player gets their Pokedex, so first encounters only start counting after that point.