Implement Battle Anim Speed Option & Update Help System
A new entry in the options menu, Battle Anim Speed, can skip the slide-in intro animation at the start of every battle and make healthbars drop much more quickly. The Help System has been updated with info on this option. Cancel was removed from the menu to make room for this, so the blue bar on top that shows controls had the AB button text changed from cancel to save, to make it more clear. This change was carried over to the Key System menu as well.

Additionally, the Help System entries for some Key System keys were updated to be more accurate.