Respawn Legendary Encounters After Hall of Fame
Legendary Pokemon, roaming Pokemon, and the Snorlax now regenerate if they weren't caught after entering the Hall of Fame.

The roamer is now released when first entering the Hall of Fame instead of after giving Celio the Sapphire. Once caught, the next roamer will be released when entering the Hall of Fame allowing for all three legendary beasts to be caught in one playthrough.

Additionally, fixed a vanilla misspelling in the Teachy TV dialogue.