Fix Underwater Bubbles Weather
The underwater bubbles weather wouldn't spawn the bubbles when diving down into a map until the map was refreshed. It was switching from a 16x32 to a 32x32 sprite, which was interfering with loading the bubble's sprites. This was due to how FR switches the player avatar's sprites, which was rewritten for the quest log. Now the player is destroyed and reinitialized in the diving state to prevent this issue.