Add Predict Battle Style to Options Menu
The battle style option now has three selections: predict, switch, and set. Predict is the old shift. Switch is shift without telling the player what is coming in next. Set is the same. The default setting for this is predict. The help menu has been updated to explain this.

Additionally, stereo is now the default sound option.