Finish Battle Tower
With the exception of the region map, the Battle Tower is done.

This commit adds door animations; fixes the elevator animation; handles Nuzlocke, No Free Heals, and Stat Calc Keys; and makes every facility class use the FRLG trainer pics if possible.

The Battle Tower will heal fainted Pokemon within it even in Nuzlocke or No Free Heals, but once the challenge is complete the party is reloaded and they will be fainted again. Stat Calc Keys are turned off after the party is backed up and the tower party's stats are recalculated to be standard.

Additionally, fixed the heal sound done in scripts so now it's correct instead of the level up sound.