Extend Safari Zone, Bond Bridge, & Berry Forest
The Safari Zone now has two new areas available only after the player becomes the Champion. These areas make the Pokemon in Emerald's Safari Zone expansion available (with the exception of Spinarak and Ledyba, who are available in Pattern Bush and are replaced with Girafarig).

Bond Bridge has been expanded to accommodate an entrance to the new Berry Forest area and the dive route to get there. Underwater Bond Bridge makes Chinchou available.

Berry Forest's extension makes the pinch berries available.

All Pokemon are spoofed to be from Emerald's equivalent locations for legality purposes.

With this commit, every Pokemon that the player can see in their Pokedex should be available, with the exception of Smeargle.