Fix Town Map Display for New Maps
Underwater Bond Bridge now displays the correct region map and position if the Town Map is opened.

The Battle Tower and Artisan Cave use the same handling as Faraway Island for when the map is opened. If I ever get the Hoenn map working, they'll be placed on it instead.