Misc Safari Zone Expansion Additions
The Safari Zone clerk now tells the player about the Safari Zone expansion the first time they try to enter the Safari Game after unlocking it. The info attendant now says there are either four or six zones depending on if the expansion is open or not.

If the expansion is not open, the Pokedex Area screen will not show Pokemon from the expansion zones as being from there.

Additionally, the Pokedex Area screen now handles Chinchou from Bond Bridge underwater.