Make Remaining Event Tickets Available
The Aurora and Eon Tickets are now obtained from getting a 28 and 56-win streak in the Battle Tower, respectively.

The Vermilion dockmaster has been completely rewritten. All event islands are now off in another menu option and checked by a function. This should now be easier to maintain in case more destinations need to be done.