Move Relearner Can Teach Moves From Pre-Evos
In what is possibly the messiest implementation in history, the Move Reminder can teach moves from pre-evos, with some caveats.

If the Pokemon has a baby form and wasn't hatched, it can't relearn those moves (Pikachu evolved from bred Pichu can relearn Charm. Caught Pikachu can't). This does not apply to Togepi. Wynaut/Wobbuffet caught on Mirage Island and Hordel's Elekid/Electabuzz from XD also count as hatched for this.

If a Pokemon shares a move with a pre-evo but is not a high enough level to learn it as its current form, it cannot relearn it (Vulpix learns Fire Spin at 41. Ninetales learns it at 45. A level 44 Ninetales cannot relearn Fire Spin).

The debug statements have been left in but commented out in case I have to go back to this.