Map & Script Four Island Base & Cable Car Stations
A new area, the base of Four Island, has been mapped and scripted. The event where the player meets the Rival on Four Island has been moved to here, as well as the items on the beach.

Both cable car stations have been mapped and scripted. They are fully functional.

Four Island has been remapped to include a cable car station and to be perched on top of the mountain. A bigger aesthetic revamp still has to be done.

Four Island now plays the Fallarbor Town music.