Five Island Aesthetic Revamp
Basically giving Verdanturf's short grass to the meadow, Mossdeep's trees in lieu of some sea rocks and for decoration on the beaches, and recolored the Five Island houses. Just enough to make it distinct.

Swapped the PMC and the house and lined up the harbor entrance with the flyspot to make it more convenient to leave Four and Five Island.

Also gave the Rocket Warehouse's door the sliding glass door sound.