Do National Dex Event & Make Project Build
The last commit won't build because of a borked include that I missed. This has been corrected.

Once the player is eligible to receive the National Pokedex, one of Oak's Aides will ambush them the next time they try to leave a Pokemon Center. If the player saves on the first floor of a Pokemon Center, becomes eligible through a Colo/XD/Box link, and reloads, it will also trigger the ambush. If the player hatches an egg whose dex entry makes them eligible in the first floor of a Pokemon Center, it will also trigger the ambush.

Considering triggering it immediately on saveload if the player reloads onto the exit mat (and therefore can leave while bypassing the trigger). While most of the time a player will become eligible for the National Dex while inside a Pokemon Center, the possibility of a hatched Egg making them eligible means the dex can't always be given immediately. While automatically doing it on saveload on the exit mat is expedient, I don't know how much it really matters given the egg situation.

The national_dex_aide's poryscript is also included in this commit for posterity.