National Dex Aide Event Fixes
The aide event now triggers immediately on saveload if the player is on a PMC exit mat and is eligible.

The aide event now checks that the player has the Extended Dex before triggering. The National Dex can now only be obtained if the player already has the Extended Dex.

Additionally, the autorun script now dismisses the map name popup, fixing a bug where the textbox summoned would move like the popup.