Do Corsola Egg Event
The first prize for showing Selphy a Pokemon is a Corsola Egg. This is implied by an addition to her team in Lost Cave and the painting in her house's new flavor text. It is generated like the Rival Egg, but has a different OT name from the player so it looks like it was traded. This makes it legal.

Additionally, fixed a vanilla tile error in Selphy's house.