Fix Freshly Caught Pokedex Screen Not Showing For Extended
When a Pokemon that is in the Extended Dex is caught, it should now show the dex entry even if the Pokemon is not in the Kanto Dex and the player does not have the National Dex. Additionally, it will now show the Extended Dex number on the first page of the Pokedex entry when shown after catching in this situation.