Make Swimmer F Accurate to Trainer Sprite
Swimmer F Land (which is apparently only used in one place in the vanilla game) now has the right hair and swimsuit color.

Swimmer F Water is now based on the land sprite instead of what is probably a Gen 2 design.

The original Swimmer F Water is still in, but unused for now.