Place Master Trainers Up to Route 4
This includes Routes 4 and 23 as well as Victory Road.

The Moltres Master is placed, and his movement type is handled, but he has no script as he's not a battleable Master Trainer.

This adds 34 Master Trainers, including the Moltres Master.

Again, dialogue for classes other than Lass/Beauty is not done.