Fix Potential Inventory Bug
The function used to save and reload the player's bag in the Cable Club did not take into account the Link Bracelet, causing bag corruption. This has been fixed.

While I'm not sure exactly when this happened to corrupt the bag in the test save, it makes sense that it happened after testing link battles.