Do Title Selection Dynamic Multichoice
The woman at Game Freak's offices in Celadon now has a dynamically-filled multichoice allowing the player to choose their title. It's actually a list menu with the arrows and everything but whatever.

Changed the Grand Master title acquisition sound to the MUS_EVOLVED fanfare to more closely match LGPE.

Also fixed an issue where title acquisition text was in red instead of dark gray, stemming from a pret correction in the color codes for the charmap.