Route 5 Day Care Man Tells What Level Deposited At
The Route 5 Day Care Man will now tell the player what level the Pokémon was when it was deposited. Since I always use this Day Care earlygame to raise Magikarp, I figured others use it in a similar manner (as a skip to get to a weak mon's evolution). This will make it easier to withdraw the Pokémon one level before its evolution, as you don't need to remember what level you deposited at.