Fix Mewtwo, Cinnabar Bill, & Four Isle Base Mapping Issues
The Mewtwo in the alternate Cerulean Cave layouts should now properly despawn after battle.

Bill in Cinnabar Island should now no longer walk through buildings when the player declines to go with him to One Island.

A tile error with a rock in the deep water of Four Island's base is now fixed.