Fix Snowflake Weather & Reflection Issue on Four Island
Four Island's occasional snow was heavy enough that it would hit the sprite limit, making the player's reflection not spawn in the lake. Reduced the snowflake count to fix this.

Additionally, when I originally "fixed" the snowflake weather, I somehow thought it was a good idea to strip out a ton of important code. This code has been restored and now the snowflakes should never start erroneously falling in a single column.