Fix Option Typo, Update FEATURES.md, Prepare Patch
Accidentally forgot the period in EXP. BAR ANIM SPEED and fixed that in both the options menu and help system.

Updated FEATURES.md with the latest changes.

Updated some screenshots and changed some pngs to gifs. The links will be broken in the thread if not updated.

Patch was made. This is a release commit.