Fix Crash While Switching Bag Pockets
The bag should no longer crash while wrapping around. There is still a bug with this where fast pocket switching will skip the Medicine pocket if switching left or the Held Items pocket if switching right. That will be fixed later, as this bug is too critical to wait to fix.

Updated the patch.

Updated Version Toggle.gif because it showed the old Key System Menu.