Make Most Warps Faster
Game Freak intentionally slowed down warps between maps with different music by artificially lengthening the fadeout time for the music. The map doesn't fade in from black until this time expires.

This commit sets it to the minimum possible time, making warps much faster.

Additionally, restored the nurse checking the Union Room for nearby players. The check for this is hidden in her heal animation. This avoids the lag inherent with entering the Pokemon Center and puts the check right before it is needed.